+++ packages/frameworks/solid/src/date-picker/stories/date-range-picker.stories.tsx
@@ -14,10 +14,12 @@ export const Basic = () => {
     <DatePicker.Root selectionMode="range">
       <DatePicker.Label>Label</DatePicker.Label>
       <DatePicker.Control>
-        <DatePicker.Input />
+        <DatePicker.Input index={0} />
+        <DatePicker.Input index={1} />
         <DatePicker.Trigger>📅</DatePicker.Trigger>
         <DatePicker.ClearTrigger>Clear</DatePicker.ClearTrigger>
       </DatePicker.Control>
+      <DatePicker.PresetTrigger value="last7Days">Last 7 days</DatePicker.PresetTrigger>
       <Portal>
         <DatePicker.Positioner>
           <DatePicker.Content>