+++ packages/frameworks/react/CHANGELOG.md
@@ -9,6 +9,7 @@ description: All notable changes to this project will be documented in this file
 ### Added
 
 - Add an optional `index` prop to the `DatePicker.Input` to support multiple inputs.
+- Add the `DatePicker.PresetTrigger` component
 - Add `open.controlled` prop in `Dialog`, `HoverCard`, `Menu`, `Popover`, `Select`, and `Tooltip` component to improve a controlled state
 
 ### Fixed